thinkAmi-sandbox/Django_separate_model_file-sample/requirements.txt
colorama==0.3.7
Django==1.9.5
factory-boy==2.7.0
fake-factory==0.5.7
py==1.4.31
pytest==2.9.1
pytest-django==2.9.1
python-dateutil==2.5.3
six==1.10.0
wheel==0.24.0
